Unicorn/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
html-parser==0.2
idna==3.4
iniconfig==2.0.0
langdetect==1.0.9
loguru==0.7.2
lxml==4.9.3
multidict==6.0.4
mysql-connector-python==8.0.33
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
ply==3.11
protobuf==3.20.3
pycparser==2.21
pyreadline==2.1
PySocks==1.7.1
pytest==7.4.3
python-docx==1.0.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==2.0.3
webdriver-manager==4.0.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.2
